Add shebang and command line parsing.
Update python code to have shebang,
accept a hand from the command-line and print the results.
Update the test_python to test the command line parsing.

Some stuff is still printed out, as opposed to being returned
when a hand of 6 is specified at the command line - so that is
not tested yet.